applications/workflows_automated/Poster/requirements.txt
Flask==2.0.1
google-api-python-client==2.31.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
tweepy==4.3.0
python-dotenv==0.19.0
schedule==1.1.0